+++ Changelog.md
@@ -3,8 +3,10 @@
 ## Unreleased
 #### 2024-XX-YY
 
+- Added Vanilo 4 support
 - Added Laravel 11 support
 - Dropped Laravel 9 support
+- Dropped Vanilo 3 support
 - Dropped PHP 8.0 & PHP 8.1 support 
 
 ## 1.1.0